// types.ts
export enum Errors {
  Database = "Database",
  Authentication = "Authentication",
}
export enum Rarity {
  Common = "Common",
  Rare = "Rare",
  Epic = "Epic",
  Dark = "Dark",
  Legendary = "Legendary",
  Mythic = "Mythic",
}
export enum Type {
  Magic = "Magic",
  Companion = "Companion",
  Summon = "Summon",
  Spell = "Spell",
}